requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
langchain==0.3.19
langchain-community==0.3.18
langchain-experimental==0.3.4
langchain-openai==0.3.6
pandas==2.2.3
tabulate==0.9.0
pydantic==2.10.6

plotly==5.3.1
kaleido==0.2.1
numpy

# File Handling
aiofiles==22.1.0
anyio==4.6.1

# Frameworks and APIs
fastapi==0.115.5
uvicorn==0.32.0
faiss-cpu
openai-agents==0.3.3
markdown
gradio==5.19.0